+++ src/components/widgets/Dialog.svelte
@@ -7,6 +7,7 @@
     import MarkupHtmlView from '../concepts/MarkupHTMLView.svelte';
     import Emoji from '@components/app/Emoji.svelte';
     import setKeyboardFocus from '@components/util/setKeyboardFocus';
+    import { clickOutside } from '@components/app/clickOutside';
 
     interface Props {
         show?: boolean;
@@ -54,24 +55,27 @@
 <!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
 <dialog
     bind:this={view}
+    use:clickOutside={() => (show = false)}
     tabindex="-1"
     onkeydown={closeable
         ? (event) => (event.key === 'Escape' ? (show = false) : undefined)
         : undefined}
 >
-    {#if closeable}
-        <div class="close">
-            <Button
-                tip={$locales.get((l) => l.ui.widget.dialog.close)}
-                action={() => (show = false)}>❌</Button
-            >
-        </div>
-    {/if}
+    <div class="container">
+        {#if closeable}
+            <div class="close">
+                <Button
+                    tip={$locales.get((l) => l.ui.widget.dialog.close)}
+                    action={() => (show = false)}>❌</Button
+                >
+            </div>
+        {/if}
 
-    <div class="content">
-        <Header>{description.header}</Header>
-        <MarkupHtmlView markup={description.explanation} />
-        {@render children?.()}
+        <div class="content">
+            <Header>{description.header}</Header>
+            <MarkupHtmlView markup={description.explanation} />
+            {@render children?.()}
+        </div>
     </div>
 </dialog>
 
@@ -79,7 +83,7 @@
     dialog {
         position: relative;
         border-radius: var(--wordplay-border-radius);
-        padding: 1em;
+        padding: 0;
         margin-left: auto;
         margin-right: auto;
         max-width: 95vw;
@@ -92,6 +96,13 @@
     dialog::backdrop {
         transition: backdrop-filter;
         backdrop-filter: blur(2px);
+    }
+
+    .container {
+        display: flex;
+        flex-direction: column;
+        gap: var(--wordplay-spacing);
+        padding: 1em;
     }
 
     .close {