+++ CHANGELOG.md
@@ -5,6 +5,16 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 
 ## 0.14.1 2024-01-11
 
+### Added
+
+- Click outside dialog to dismiss it.
+- Fixed #603 Machine translation of tutorials.
+
+### Fixed
+
+- Tutorial typos
+- More precise empty input lists in locale schema.
+
 ### Maintenence
 
 - Cleaned up the supported locales code.